Upstream version 2.3

+++ test/test_collection.py
@@ -44,7 +44,7 @@ from pymongo.errors import (ConfigurationError,
                             OperationFailure,
                             TimeoutError)
 from test.test_connection import get_connection
-from test.utils import joinall
+from test.utils import is_mongos, joinall
 from test import (qcheck,
                   version)
 
@@ -62,6 +62,7 @@ class TestCollection(unittest.TestCase):
         self.db = self.connection.pymongo_test
 
     def tearDown(self):
+        self.db.drop_collection("test_large_limit")
         self.db = None
         self.connection = None
 
@@ -185,22 +186,39 @@ class TestCollection(unittest.TestCase):
 
         db.test.drop_index("goodbye_1")
         self.assertEqual("goodbye_1",
-                         db.test.ensure_index("goodbye", ttl=1))
+                         db.test.ensure_index("goodbye", cache_for=1))
         time.sleep(1.1)
         self.assertEqual("goodbye_1",
                          db.test.ensure_index("goodbye"))
 
         db.test.drop_index("goodbye_1")
         self.assertEqual("goodbye_1",
-                         db.test.create_index("goodbye", ttl=1))
+                         db.test.create_index("goodbye", cache_for=1))
         time.sleep(1.1)
         self.assertEqual("goodbye_1",
                          db.test.ensure_index("goodbye"))
         # Make sure the expiration time is updated.
         self.assertEqual(None,
                          db.test.ensure_index("goodbye"))
+
         # Clean up indexes for later tests
         db.test.drop_indexes()
+
+    def test_deprecated_ttl_index_kwarg(self):
+        db = self.db
+
+        # In Python 2.6+ we could use the catch_warnings context
+        # manager to test this warning nicely. As we can't do that
+        # we must test raising errors before the ignore filter is applied.
+        warnings.simplefilter("error", DeprecationWarning)
+        self.assertRaises(DeprecationWarning, lambda:
+                        db.test.ensure_index("goodbye", ttl=10))
+        warnings.resetwarnings()
+        warnings.simplefilter("ignore")
+
+        self.assertEqual("goodbye_1",
+                         db.test.ensure_index("goodbye", ttl=10))
+        self.assertEqual(None, db.test.ensure_index("goodbye"))
 
     def test_ensure_unique_index_threaded(self):
         coll = self.db.test_unique_threaded
@@ -282,15 +300,24 @@ class TestCollection(unittest.TestCase):
         db.test.create_index("who")
         db.test.create_index("when")
         info = db.test.index_information()
+
+        def check_result(result):
+            self.assertEqual(4, result['nIndexes'])
+            self.assertEqual(4, result['nIndexesWas'])
+            indexes = result['indexes']
+            names = [idx['name'] for idx in indexes]
+            for name in names:
+                self.assertTrue(name in info)
+            for key in info:
+                self.assertTrue(key in names)
+
         reindexed = db.test.reindex()
-        self.assertEqual(4, reindexed['nIndexes'])
-        self.assertEqual(4, reindexed['nIndexesWas'])
-        indexes = reindexed['indexes']
-        names = [idx['name'] for idx in indexes]
-        for name in names:
-            self.assertTrue(name in info)
-        for key in info:
-            self.assertTrue(key in names)
+        if 'raw' in reindexed:
+            # mongos
+            for result in reindexed['raw'].itervalues():
+                check_result(result)
+        else:
+            check_result(reindexed)
 
     def test_index_info(self):
         db = self.db
@@ -324,6 +351,8 @@ class TestCollection(unittest.TestCase):
         self.assertEqual([('loc', '2d')], index_info['key'])
 
     def test_index_haystack(self):
+        if is_mongos(self.db.connection):
+            raise SkipTest("geoSearch is not supported by mongos")
         db = self.db
         db.test.drop_indexes()
         db.test.remove()
@@ -633,17 +662,17 @@ class TestCollection(unittest.TestCase):
 
     def test_invalid_key_names(self):
         db = self.db
-        db.test.remove({})
+        db.test.drop()
 
-        db.test.insert({"hello": "world"})
-        db.test.insert({"hello": {"hello": "world"}})
+        db.test.insert({"hello": "world"}, safe=True)
+        db.test.insert({"hello": {"hello": "world"}}, safe=True)
 
         self.assertRaises(InvalidDocument, db.test.insert, {"$hello": "world"})
         self.assertRaises(InvalidDocument, db.test.insert,
                           {"hello": {"$hello": "world"}})
 
-        db.test.insert({"he$llo": "world"})
-        db.test.insert({"hello": {"hello$": "world"}})
+        db.test.insert({"he$llo": "world"}, safe=True)
+        db.test.insert({"hello": {"hello$": "world"}}, safe=True)
 
         self.assertRaises(InvalidDocument, db.test.insert,
                           {".hello": "world"})
@@ -666,7 +695,7 @@ class TestCollection(unittest.TestCase):
         doc1 = {"hello": u"world"}
         doc2 = {"hello": u"mike"}
         self.assertEqual(db.test.find().count(), 0)
-        ids = db.test.insert([doc1, doc2])
+        ids = db.test.insert([doc1, doc2], safe=True)
         self.assertEqual(db.test.find().count(), 2)
         self.assertEqual(doc1, db.test.find_one({"hello": u"world"}))
         self.assertEqual(doc2, db.test.find_one({"hello": u"mike"}))
@@ -684,6 +713,7 @@ class TestCollection(unittest.TestCase):
     def test_insert_multiple_with_duplicate(self):
         db = self.db
         db.drop_collection("test")
+        db.safe = True
         db.test.ensure_index([('i', ASCENDING)], unique=True)
 
         # No error
@@ -691,12 +721,39 @@ class TestCollection(unittest.TestCase):
         db.test.remove()
 
         # No error
-        db.test.insert([{'i': 1}] * 2)
+        db.test.insert([{'i': 1}] * 2, safe=False)
         self.assertEqual(1, db.test.count())
 
         self.assertRaises(
             DuplicateKeyError,
+            lambda: db.test.insert([{'i': 2}] * 2),
+        )
+
+        # Test based on default setting as False
+        db.drop_collection("test")
+        db.safe = False
+        db.test.ensure_index([('i', ASCENDING)], unique=True)
+
+        # No error
+        db.test.insert([{'i': 1}] * 2)
+        self.assertEqual(1, db.test.count())
+
+        # Implied safe
+        self.assertRaises(
+            DuplicateKeyError,
+            lambda: db.test.insert([{'i': 2}] * 2, w=1),
+        )
+
+        # Explicit safe
+        self.assertRaises(
+            DuplicateKeyError,
             lambda: db.test.insert([{'i': 2}] * 2, safe=True),
+        )
+
+        # Misconfigured value for safe
+        self.assertRaises(
+            TypeError,
+            lambda: db.test.insert([{'i': 2}] * 2, safe=1),
         )
 
     def test_insert_iterables(self):
@@ -823,7 +880,7 @@ class TestCollection(unittest.TestCase):
     def test_continue_on_error(self):
         db = self.db
         if not version.at_least(db.connection, (1, 9, 1)):
-            raise SkipTest()
+            raise SkipTest("continue_on_error requires MongoDB >= 1.9.1")
 
         db.drop_collection("test")
         oid = db.test.insert({"one": 1})
@@ -909,7 +966,7 @@ class TestCollection(unittest.TestCase):
     def test_multi_update(self):
         db = self.db
         if not version.at_least(db.connection, (1, 1, 3, -1)):
-            raise SkipTest()
+            raise SkipTest("multi-update requires MongoDB >= 1.1.3")
 
         db.drop_collection("test")
 
@@ -1003,7 +1060,7 @@ class TestCollection(unittest.TestCase):
 
     def test_last_error_options(self):
         if not version.at_least(self.connection, (1, 5, 1)):
-            raise SkipTest()
+            raise SkipTest("getLastError options require MongoDB >= 1.5.1")
 
         # XXX: Fix this if we ever have a replica set unittest env.
         # mongo >=1.7.6 errors with 'norepl' when w=2+
@@ -1016,7 +1073,7 @@ class TestCollection(unittest.TestCase):
             self.assertRaises(TimeoutError, self.db.test.update,
                               {"x": 1}, {"y": 2}, w=2, wtimeout=1)
             self.assertRaises(TimeoutError, self.db.test.remove,
-                              {"x": 1}, {"y": 2}, w=2, wtimeout=1)
+                              {"x": 1}, w=2, wtimeout=1)
 
         self.db.test.save({"x": 1}, w=1, wtimeout=1)
         self.db.test.insert({"x": 1}, w=1, wtimeout=1)
@@ -1045,6 +1102,21 @@ class TestCollection(unittest.TestCase):
         db.test.save({'foo': 'baz'})
         self.assertEqual(db.test.find({'foo': 'bar'}).count(), 1)
         self.assertEqual(db.test.find({'foo': re.compile(r'ba.*')}).count(), 2)
+
+    def test_aggregate(self):
+        if not version.at_least(self.db.connection, (2, 1, 0)):
+            raise SkipTest("The aggregate command requires MongoDB >= 2.1.0")
+        db = self.db
+        db.drop_collection("test")
+        db.test.save({'foo': [1, 2]})
+
+        self.assertRaises(TypeError, db.test.aggregate, "wow")
+
+        pipeline = {"$project": {"_id": False, "foo": True}}
+        expected = {'ok': 1.0, 'result': [{'foo': [1, 2]}]}
+        self.assertEqual(expected, db.test.aggregate(pipeline))
+        self.assertEqual(expected, db.test.aggregate([pipeline]))
+        self.assertEqual(expected, db.test.aggregate((pipeline,)))
 
     def test_group(self):
         db = self.db
@@ -1154,22 +1226,26 @@ class TestCollection(unittest.TestCase):
 
     def test_large_limit(self):
         db = self.db
-        db.drop_collection("test")
-        db.test.create_index([('x', 1)])
+        db.drop_collection("test_large_limit")
+        db.test_large_limit.create_index([('x', 1)])
 
         for i in range(2000):
-            db.test.insert({"x": i, "y": "mongomongo" * 1000}, safe=True)
+            doc = {"x": i, "y": "mongomongo" * 1000}
+            db.test_large_limit.insert(doc, safe=True)
 
         # Wait for insert to complete; often mysteriously failing in Jenkins
         st = time.time()
-        while len(list(db.test.find())) < 2000 and time.time() - st < 30:
+        while (
+            len(list(db.test_large_limit.find())) < 2000
+            and time.time() - st < 30
+        ):
             time.sleep(1)
 
-        self.assertEqual(2000, len(list(db.test.find())))
+        self.assertEqual(2000, len(list(db.test_large_limit.find())))
 
         i = 0
         y = 0
-        for doc in db.test.find(limit=1900).sort([('x', 1)]):
+        for doc in db.test_large_limit.find(limit=1900).sort([('x', 1)]):
             i += 1
             y += doc["x"]
 
@@ -1330,7 +1406,7 @@ class TestCollection(unittest.TestCase):
 
     def test_distinct(self):
         if not version.at_least(self.db.connection, (1, 1)):
-            raise SkipTest()
+            raise SkipTest("distinct command requires MongoDB >= 1.1")
 
         self.db.drop_collection("test")
 
@@ -1410,7 +1486,7 @@ class TestCollection(unittest.TestCase):
 
     def test_map_reduce(self):
         if not version.at_least(self.db.connection, (1, 1, 1)):
-            raise SkipTest()
+            raise SkipTest("mapReduce command requires MongoDB >= 1.1.1")
 
         db = self.db
         db.drop_collection("test")
@@ -1465,14 +1541,18 @@ class TestCollection(unittest.TestCase):
             self.assertEqual(2, result.find_one({"_id": "dog"})["value"])
             self.assertEqual(1, result.find_one({"_id": "mouse"})["value"])
 
-            result = db.test.map_reduce(map, reduce,
-                                        out=SON([('replace', 'mrunittests'),
-                                                 ('db', 'mrtestdb')
-                                                ]))
-            self.assertEqual(3, result.find_one({"_id": "cat"})["value"])
-            self.assertEqual(2, result.find_one({"_id": "dog"})["value"])
-            self.assertEqual(1, result.find_one({"_id": "mouse"})["value"])
-            self.connection.mrtestdb.mrunittests.drop()
+            if (is_mongos(self.db.connection)
+                and not version.at_least(self.db.connection, (2, 1, 2))):
+                pass
+            else:
+                result = db.test.map_reduce(map, reduce,
+                                            out=SON([('replace', 'mrunittests'),
+                                                     ('db', 'mrtestdb')
+                                                    ]))
+                self.assertEqual(3, result.find_one({"_id": "cat"})["value"])
+                self.assertEqual(2, result.find_one({"_id": "dog"})["value"])
+                self.assertEqual(1, result.find_one({"_id": "mouse"})["value"])
+                self.connection.mrtestdb.mrunittests.drop()
 
         full_result = db.test.map_reduce(map, reduce,
                                          out='mrunittests', full_response=True)
@@ -1559,8 +1639,15 @@ class TestCollection(unittest.TestCase):
 
         self.assertEqual(None,
                          c.find_and_modify({'_id': 1}, {'$inc': {'i': 1}}))
-        self.assertEqual({}, c.find_and_modify({'_id': 1}, {'$inc': {'i': 1}},
-                                               upsert=True))
+        # The return value changed in 2.1.2. See SERVER-6226.
+        if version.at_least(self.db.connection, (2, 1, 2)):
+            self.assertEqual(None, c.find_and_modify({'_id': 1},
+                                                     {'$inc': {'i': 1}},
+                                                     upsert=True))
+        else:
+            self.assertEqual({}, c.find_and_modify({'_id': 1},
+                                                   {'$inc': {'i': 1}},
+                                                   upsert=True))
         self.assertEqual({'_id': 1, 'i': 2},
                          c.find_and_modify({'_id': 1}, {'$inc': {'i': 1}},
                                            upsert=True, new=True))
@@ -1585,7 +1672,7 @@ class TestCollection(unittest.TestCase):
 
     def test_find_with_nested(self):
         if not version.at_least(self.db.connection, (2, 0, 0)):
-            raise SkipTest()
+            raise SkipTest("nested $and and $or requires MongoDB >= 2.0")
         c = self.db.test
         c.drop()
         c.insert([{'i': i} for i in range(5)])  # [0, 1, 2, 3, 4]
@@ -1653,7 +1740,7 @@ class TestCollection(unittest.TestCase):
 
     def test_uuid_subtype(self):
         if not have_uuid:
-            raise SkipTest()
+            raise SkipTest("No uuid module")
 
         coll = self.connection.pymongo_test.uuid
         coll.drop()
@@ -1663,7 +1750,7 @@ class TestCollection(unittest.TestCase):
 
         # Test property
         self.assertEqual(OLD_UUID_SUBTYPE, coll.uuid_subtype)
-        self.assertRaises(ConfigurationError, change_subtype, coll, 5)
+        self.assertRaises(ConfigurationError, change_subtype, coll, 7)
         self.assertRaises(ConfigurationError, change_subtype, coll, 2)
 
         # Test basic query
@@ -1759,16 +1846,18 @@ class TestCollection(unittest.TestCase):
 
         coll.uuid_subtype = UUID_SUBTYPE
         q = {"_id": uu}
-        result = coll.inline_map_reduce(map, reduce, query=q)
-        self.assertEqual([], result)
+        if version.at_least(self.db.connection, (1, 7, 4)):
+            result = coll.inline_map_reduce(map, reduce, query=q)
+            self.assertEqual([], result)
 
         result = coll.map_reduce(map, reduce, "results", query=q)
         self.assertEqual(0, db.results.count())
 
         coll.uuid_subtype = OLD_UUID_SUBTYPE
         q = {"_id": uu}
-        result = coll.inline_map_reduce(map, reduce, query=q)
-        self.assertEqual(2, len(result))
+        if version.at_least(self.db.connection, (1, 7, 4)):
+            result = coll.inline_map_reduce(map, reduce, query=q)
+            self.assertEqual(2, len(result))
 
         result = coll.map_reduce(map, reduce, "results", query=q)
         self.assertEqual(2, db.results.count())
